transactions_backend/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
psycopg2==2.9.9
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.9.0
tzdata==2024.1
django-cors-headers==4.3.1
drf-yasg==1.21.7
